+++ app/src/main/res/layout/app_bar_navigation.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context="jp.numero.designdemoapplication.MainActivity">
+    tools:context="jp.numero.designdemoapplication.NavigationActivity">
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
@@ -21,7 +21,7 @@
 
     </android.support.design.widget.AppBarLayout>
 
-    <include layout="@layout/content_main" />
+    <include layout="@layout/content_navigation" />
 
     <android.support.design.widget.FloatingActionButton
         android:id="@+id/fab"
