+++ app/src/main/res/layout/content_navigation.xml
@@ -9,18 +9,11 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="jp.numero.designdemoapplication.MainActivity"
-    tools:showIn="@layout/activity_main">
+    tools:context="jp.numero.designdemoapplication.NavigationActivity"
+    tools:showIn="@layout/app_bar_navigation">
 
-    <ScrollView
-        android:layout_width="match_parent"
-        android:layout_height="match_parent">
-        <Button
-            android:id="@+id/navigation_view"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:text="NavigationView"
-            android:foreground="?android:attr/selectableItemBackground"/>
-
-    </ScrollView>
+    <TextView
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="Hello World!" />
 </RelativeLayout>
